Refactor date converter into utils class
