+++ rdt-app/src/main/java/io/ona/rdt_app/util/Utils.java
@@ -1,8 +1,12 @@
 package io.ona.rdt_app.util;
 
+import org.apache.commons.lang3.StringUtils;
 import org.smartregister.job.SyncServiceJob;
 
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Date;
 
 import io.ona.rdt_app.BuildConfig;
 import io.ona.rdt_app.application.RDTApplication;
@@ -53,5 +57,14 @@ public class Utils {
             minutes = (long) Math.ceil(value / 3);
         }
         return minutes;
+    }
+
+    public static Date convertDate(String dateStr, String format) throws ParseException {
+        if (StringUtils.isEmpty(dateStr)) {
+            return null;
+        }
+
+        SimpleDateFormat simpleDateFormat = new SimpleDateFormat(format);
+        return simpleDateFormat.parse(dateStr);
     }
 }
